backProp tests and segregating old tests
