--- tests/autoBP.exe.stackdump
@@ -1,18 +1,0 @@
-Exception: STATUS_ACCESS_VIOLATION at rip=00100440B54
-rax=0000000000000000 rbx=0000000000000002 rcx=00000000FFFFC9B8
-rdx=0000000000000000 rsi=0000000180252AA0 rdi=0000000000000000
-r8 =00000000FFFFC98C r9 =000000018013FF20 r10=0000000100000000
-r11=00000003FEBE1D92 r12=0000000000000000 r13=00000000FFFFCC56
-r14=0000000000000000 r15=000000018021E3F0
-rbp=00000000FFFFC970 rsp=00000000FFFFC950
-program=C:\Users\J. Patrick Taggart\Documents\Work\WorkStudy\Feat\feat\tests\autoBP.exe, pid 16616, thread main
-cs=0033 ds=002B es=002B fs=0053 gs=002B ss=002B
-Stack trace:
-Frame        Function    Args
-000FFFFC970  00100440B54 (00000000000, 003FEC3E020, 003FEC3FF60, 00180252AA0)
-000FFFFCA00  0010040870E (000FFFFCB80, 001801C6C30, 000FFFFCB80, 003FEC4B8F8)
-000FFFFCA70  00100406B09 (00000000020, FF0700010302FF00, 00180047941, 00000000000)
-000FFFFCCB0  001800479B2 (00000000000, 00000000000, 00000000000, 00000000000)
-00000000000  00180045653 (00000000000, 00000000000, 00000000000, 00000000000)
-000FFFFFFF0  00180045704 (00000000000, 00000000000, 00000000000, 00000000000)
-End of stack trace
